% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2adj.R
\name{data2adj}
\alias{data2adj}
\title{construct adjacency matrix from data}
\usage{
data2adj(dat)
}
\arguments{
\item{dat}{data to cluster, either a dataframe or matrix (rows form the observations)}
}
\value{
sparse adjacency matrix
}
\description{
construct adjacency matrix from data
}
\examples{
W <- data2adj(seeds)
}
