% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optcluster.R
\name{optcluster}
\alias{optcluster}
\title{find the optimal clustering}
\usage{
optcluster(X, C)
}
\arguments{
\item{X}{data matrix, rows are the vertices to be clustered}

\item{C}{matrix, each column is one clustering result}
}
\value{
optimal clustering vector according to the Calinski-Harabasz index
}
\description{
find the optimal clustering
}
