% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_kmeans.R
\name{rep_kmeans}
\alias{rep_kmeans}
\title{rep_kmeans}
\usage{
rep_kmeans(Kd, k, n = 100)
}
\arguments{
\item{Kd}{Diffusion kernel matrix.}

\item{k}{Number of clusters to determine.}

\item{n}{Number of repeated kmeans clusterings, default is 100}
}
\value{
numeric matrix of n cluster classifications.
Each clustering is represented by one column of the matrix.
}
\description{
Repeats kmeans clusterings on a diffusion kernel matrix.
}
