% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\docType{data}
\name{seeds}
\alias{seeds}
\title{Seeds data from the UCI data repository}
\format{
\subsection{\code{seeds}}{

A data frame with 210 rows and 8 columns and measurements of 210 wheat kernels.
\describe{
\item{area}{area of kernel}
\item{perimeter}{perimeter of kernel}
\item{compactness}{4*pi * area / (perimeter^2)}
\item{length}{length of kernel}
\item{width}{width of kernel}
\item{asymmetry}{asymmetry coefficient}
\item{groove}{length of kernel groove}
\item{variety}{variety of wheat}
...
}
}
}
\source{
\url{https://archive.ics.uci.edu/dataset/236/seeds}
}
\usage{
seeds
}
\description{
A small test data for clustering ...
}
\keyword{datasets}
