% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prwdkc.R
\name{prwdkc}
\alias{prwdkc}
\title{prwdk clustering function}
\usage{
prwdkc(W, k, nu, ld)
}
\arguments{
\item{W}{adjacency matrix (or Matrix)}

\item{k}{number of clusters to determine}

\item{nu}{start distribution on the vertices}

\item{ld}{2^ld is stop time for determining clusters}
}
\value{
clustering as vector indicating per vertex the assigned cluster number
}
\description{
prwdk clustering function
}
